refactor: refactor worker and task metrics to use int64 types (#136)
- Change `BusyWorkers` type from `uint64` to `int64`
- Add `CompletedTasks` method to `Metric` interface and implementation
- Update `IncBusyWorker` and `DecBusyWorker` to use `atomic.AddInt64`
- Modify test assertions to use `uint64` and `int64` types
- Change `defaultWorkerCount` and related functions from `int` to `int64`
- Update `NewPool` and related functions to use `int64` for size and worker count
- Add `CompletedTasks` method to `Queue` and update related methods to use `uint64` and `int64` types

+++ pool.go
@@ -1,7 +1,7 @@
 package queue
 
 // NewPool initializes a new pool
-func NewPool(size int, opts ...Option) *Queue {
+func NewPool(size int64, opts ...Option) *Queue {
 	o := []Option{
 		WithWorkerCount(size),
 		WithWorker(NewRing(opts...)),
